chore: convert `detectCurrentPlatform` to CJS
