--- .eslintignore
@@ -1,6 +1,0 @@
-node_modules
-
-cache
-build
-dist
-out
