+++ lib/cli.cjs
@@ -4,7 +4,7 @@ const yargs = require("yargs/yargs");
 const { hideBin } = require("yargs/helpers");
 
 const { nwbuild } = require("./index.cjs");
-const { detectCurrentPlatform } = require("../dist/index.cjs");
+const detectCurrentPlatform = require("./util/detectCurrentPlatform.cjs");
 
 const cli = yargs(hideBin(process.argv))
   .version(false)
