+++ package.json
@@ -29,7 +29,6 @@
   "bin": {
     "nwbuild": "./lib/cli.cjs"
   },
-  "type": "module",
   "files": [
     "bin",
     "lib",
@@ -45,7 +44,6 @@
     "lint": "eslint --config=cfg/eslint.config.cjs src test",
     "lint:fix": "eslint --config=cfg/eslint.config.cjs --fix src test",
     "test:prep": "esbuild src/index.js --bundle --platform=node --outfile=./dist/index.cjs",
-    "test:unit": "node --test test/unit/detectCurrentPlatform.js ",
     "test:tape": "tape './test/*.cjs'",
     "test:sanity": "node --test-reporter=spec --test e2e/test.js",
     "demo": "npm run test:prep && npm link nw-builder && cd test && node demo.cjs"
