+++ src/index.js
@@ -1,7 +1,5 @@
 import Options from "./constants/Options";
-import Platform from "./constants/Platform";
 import Platforms from "./constants/Platforms";
-import detectCurrentPlatform from "./utilities/detectCurrentPlatform";
 
 import { get } from "./get.js";
 import { log, setLogLevel } from "./log.js";
@@ -10,9 +8,7 @@ import { isCached, parse, getReleaseInfo, validate } from "./util.js";
 
 export {
   Options,
-  Platform,
   Platforms,
-  detectCurrentPlatform,
   get,
   getReleaseInfo,
   isCached,
